requirements.txt
fastapi>=0.70.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8
opencv-python>=4.5
numpy>=1.19
tensorflow>=2.6
python-jose>=3.3.0